requirements.txt
ag2==0.9.9
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asyncer==0.0.8
attrs==25.3.0
autogen==0.9.9
autogen-agentchat==0.7.4
autogen-core==0.7.4
beautifulsoup4==4.13.5
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
chardet==3.0.4
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cssselect==1.3.0
dataclasses-json==0.6.7
defusedxml==0.7.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
duckduckgo-search==3.8.5
faiss-cpu==1.12.0
fastapi==0.116.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.19.1
filetype==1.2.0
Flask==3.1.2
Flask-SocketIO==5.5.1
fonttools==4.59.2
fpdf2==2.8.4
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.111.0
google-cloud-bigquery==3.37.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.35.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-google-genai==2.0.10
langchain-openai==0.3.32
langchain-text-splitters==0.3.11
langdetect==1.0.9
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.21
llvmlite==0.44.0
lxml==6.0.1
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
marshmallow==3.26.1
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
narwhals==2.3.0
networkx==3.5
newspaper3k==0.2.8
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
openai==1.102.0
openai-whisper==20250625
opentelemetry-api==1.36.0
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyautogen==0.10.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.1
reportlab==4.4.3
requests==2.32.5
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
sgmllib3k==1.0.0
shapely==2.1.1
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
source==1.2.0
SQLAlchemy==2.0.43
starlette==0.47.3
streamlit==1.49.1
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.22.0
toml==0.10.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
